Fix integration tests date bug and refactor DateFormats
