+++ server/constants/dateFormats.ts
@@ -1,4 +1,5 @@
 // eslint-disable-next-line import/prefer-default-export, no-shadow
 export enum DateFormats {
   ISO_DATE = 'yyyy-MM-dd',
+  PRETTY_DATE = 'EEEE d MMMM yyyy',
 }
