+++ server/services/visitSessionsService.ts
@@ -1,7 +1,7 @@
 import { addDays, eachDayOfInterval, format } from 'date-fns'
 import { HmppsAuthClient, OrchestrationApiClient, RestClientBuilder } from '../data'
 import { AvailableVisitSessionDto } from '../data/orchestrationApiTypes'
-import { DateFormats } from '../utils/constants'
+import { DateFormats } from '../constants/dateFormats'
 
 export type SessionRestriction = AvailableVisitSessionDto['sessionRestriction']
